@@ -2,10 +2,10 @@ package com.ekdorn.stealapeak.database;
 
 import java.util.List;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
 
 @Dao
 public interface MessageDao {
